Separacion de wrapper ReadOnly WriteOnly
