+++ src/infrastructure/InterviewAPI.Persistence/Abstractions/Crud/ICrudRepository.cs
@@ -3,12 +3,14 @@ using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 
-namespace InterviewAPI.Persistence.Abstractions.Queries
+namespace InterviewAPI.Persistence.Abstractions.Crud
 {
-    public interface IReadOnlyRepository<T>
+    public interface ICrudRepository<T>
     {
         public Task<List<T>> GetAll();
         public Task<List<T>> GetByCondition(Expression<Func<T, bool>> expression);
-        
+        public void Create(T entity);
+        public void Update(T entity);
+        public void Delete(T entity);
     }
 }
