+++ src/infrastructure/InterviewAPI.Persistence/Abstractions/Crud/IInterviewerCrudRepository.cs
@@ -1,6 +1,6 @@
 using InterviewAPI.Entities.Models;
 
-namespace InterviewAPI.Persistence.Abstractions.Commands
+namespace InterviewAPI.Persistence.Abstractions.Crud
 {
     public interface IInterviewerCrudRepository : ICrudRepository<Interviewer>
     {
