+++ src/infrastructure/InterviewAPI.Persistence/Abstractions/ReadOnly/IIntervieweeReadOnlyRepository.cs
@@ -1,6 +1,6 @@
 using InterviewAPI.Entities.Models;
 
-namespace InterviewAPI.Persistence.Abstractions.Queries
+namespace InterviewAPI.Persistence.Abstractions.ReadOnly
 {
     public interface IIntervieweeReadOnlyRepository : IReadOnlyRepository<Interviewee>
     {
