+++ src/infrastructure/InterviewAPI.Persistence/Abstractions/ReadOnly/IReadOnlyRepository.cs
@@ -3,14 +3,12 @@ using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 
-namespace InterviewAPI.Persistence.Abstractions.Commands
+namespace InterviewAPI.Persistence.Abstractions.ReadOnly
 {
-    public interface ICrudRepository<T>
+    public interface IReadOnlyRepository<T>
     {
         public Task<List<T>> GetAll();
         public Task<List<T>> GetByCondition(Expression<Func<T, bool>> expression);
-        public void Create(T entity);
-        public void Update(T entity);
-        public void Delete(T entity);
+        
     }
 }
