+++ src/infrastructure/InterviewAPI.Persistence/Repositories/Crud/InterviewCrudRepository.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 using InterviewAPI.Entities.Models;
-using InterviewAPI.Persistence.Abstractions.Commands;
+using InterviewAPI.Persistence.Abstractions.Crud;
 using InterviewAPI.Persistence.Context;
 using Microsoft.EntityFrameworkCore;
 
-namespace InterviewAPI.Persistence.Repositories.Commands
+namespace InterviewAPI.Persistence.Repositories.Crud
 {
     public class InterviewCrudRepository : CrudRepository<Interview>, IInterviewCrudRepository
     {
