+++ src/infrastructure/InterviewAPI.Persistence/Repositories/Crud/InterviewerCrudRepository.cs
@@ -4,32 +4,36 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 using InterviewAPI.Entities.Models;
-using InterviewAPI.Persistence.Abstractions.Queries;
+using InterviewAPI.Persistence.Abstractions.Crud;
 using InterviewAPI.Persistence.Context;
 using Microsoft.EntityFrameworkCore;
 
-namespace InterviewAPI.Persistence.Repositories.Queries
+namespace InterviewAPI.Persistence.Repositories.Crud
 {
-    public class InterviewerReadOnlyRepository : ReadOnlyRepository<Interviewer>, IInterviewerReadOnlyRepository
+    public class InterviewerCrudRepository : CrudRepository<Interviewer>, IInterviewerCrudRepository
     {
-        public InterviewerReadOnlyRepository(InterviewContext interviewContext) : base(interviewContext)
+        public InterviewerCrudRepository(InterviewContext interviewContext) : base(interviewContext)
         {
         }
+
         public override async Task<List<Interviewer>> GetAll()
         {
             return await InterviewContext.Set<Interviewer>()
-                .Include(interviewer => interviewer.Interviews)
-                .AsNoTracking()
+                .Include(i => i.Interviews)
                 .ToListAsync();
         }
 
         public override async Task<List<Interviewer>> GetByCondition(Expression<Func<Interviewer, bool>> expression)
         {
-            return await InterviewContext.Set<Interviewer>()
-                .Include(interviewer => interviewer.Interviews)
+            return await InterviewContext.Interviewers
+                .Include(i => i.Interviews)
                 .Where(expression)
-                .AsNoTracking()
                 .ToListAsync();
+        }
+
+        public override void Create(Interviewer entity)
+        {
+            InterviewContext.Set<Interviewer>().Add(entity);
         }
     }
 }
