+++ src/infrastructure/InterviewAPI.Persistence/Repositories/ReadOnly/IntervieweeReadOnlyRepository.cs
@@ -1,8 +1,8 @@
 using InterviewAPI.Entities.Models;
-using InterviewAPI.Persistence.Abstractions.Queries;
+using InterviewAPI.Persistence.Abstractions.ReadOnly;
 using InterviewAPI.Persistence.Context;
 
-namespace InterviewAPI.Persistence.Repositories.Queries
+namespace InterviewAPI.Persistence.Repositories.ReadOnly
 {
     public class IntervieweeReadOnlyRepository : ReadOnlyRepository<Interviewee>, IIntervieweeReadOnlyRepository
     {
