+++ src/infrastructure/InterviewAPI.Persistence/Repositories/ReadOnly/ReadOnlyRepository.cs
@@ -3,11 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
-using InterviewAPI.Persistence.Abstractions.Queries;
+using InterviewAPI.Persistence.Abstractions.ReadOnly;
 using InterviewAPI.Persistence.Context;
 using Microsoft.EntityFrameworkCore;
 
-namespace InterviewAPI.Persistence.Repositories.Queries
+namespace InterviewAPI.Persistence.Repositories.ReadOnly
 {
     public class ReadOnlyRepository<T> : IReadOnlyRepository<T> where T : class
     {
