+++ src/infrastructure/InterviewAPI.Persistence/Repositories/RepositoryWrapper.cs
@@ -1,8 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using InterviewAPI.Persistence.Abstractions;
-using InterviewAPI.Persistence.Abstractions.Commands;
-using InterviewAPI.Persistence.Abstractions.Queries;
+using InterviewAPI.Persistence.Abstractions.Crud;
 using InterviewAPI.Persistence.Context;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -18,9 +17,6 @@ namespace InterviewAPI.Persistence.Repositories
         public IIntervieweeCrudRepository IntervieweeRepository => _serviceProvider.GetService<IIntervieweeCrudRepository>();
         public IInterviewerCrudRepository InterviewerRepository => _serviceProvider.GetService<IInterviewerCrudRepository>();
 
-        public IInterviewerReadOnlyRepository InterviewerReadOnlyRepository => _serviceProvider.GetService<IInterviewerReadOnlyRepository>();
-        public IInterviewReadOnlyRepository InterviewReadOnlyRepository => _serviceProvider.GetService<IInterviewReadOnlyRepository>();
-        public IIntervieweeReadOnlyRepository IntervieweeReadOnlyRepository => _serviceProvider.GetService<IIntervieweeReadOnlyRepository>();
 
         public async Task Save()
         {
